[MODI-77] FEAT: 파티 상세 정보 조회 API 추가
- api 추가 이외에 아래 변경사항이 있습니다
  - PartyResponse dto를 PartyBriefResponse로 변경
  - Party 엔티티에서 Member, PartyRule로 양방향 연관관계 설정
